[K/N] Use ptr-sized types in objc_firendly_dealloc tests
Merge-request: KT-MR-11525
Merged-by: Alexey Glushko <aleksei.glushko@jetbrains.com>

+++ kotlin-native/backend.native/tests/interop/objc/objc_friendly_dealloc/objclib.h
@@ -6,6 +6,6 @@
 -(uintptr_t)identity;
 @end
 
-void retain(uint64_t);
-void release(uint64_t);
-void autorelease(uint64_t);
+void retain(uintptr_t);
+void release(uintptr_t);
+void autorelease(uintptr_t);
